pyproject.toml
[project]
name = "python-pdf-llm"
version = "0.1.0"
description = ""
authors = [
    {name = "Albert Kim",email = "albert275@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.59.9,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pymupdf (>=1.25.2,<2.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
